requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
jinja2>=3.0.0
psycopg2-binary>=2.9.0  # PostgreSQL support (optional)
